+++ test/python/test_qiskit_backend.py
@@ -260,6 +260,7 @@ def circuit() -> QuantumCircuit:
     return circuit
 
 
+@pytest.mark.iqm
 def test_iqm_backend(circuit: QuantumCircuit, backend: IQMBackend) -> None:
     """Test the execution of a simple Bell state circuit."""
     circuit.measure_all()
@@ -269,6 +270,7 @@ def test_iqm_backend(circuit: QuantumCircuit, backend: IQMBackend) -> None:
     assert sum(counts.values()) == 8
 
 
+@pytest.mark.iqm
 def test_iqm_backend_sampler(circuit: QuantumCircuit, backend: IQMBackend) -> None:
     """The bound sampler should execute a simple circuit on the live IQM backend."""
     circuit.measure_all()
@@ -278,6 +280,7 @@ def test_iqm_backend_sampler(circuit: QuantumCircuit, backend: IQMBackend) -> No
     assert sum(counts.values()) == 8
 
 
+@pytest.mark.iqm
 def test_iqm_backend_estimator(circuit: QuantumCircuit, backend: IQMBackend) -> None:
     """The bound estimator should execute a simple observable on the live IQM backend."""
     observable = SparsePauliOp("Z" * backend.num_qubits)